backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]
python-multipart
pymupdf
python-docx
weasyprint
sentence-transformers
langchain>=0.2.0
langchain-anthropic
langchain-openai
langchain-google-genai
cryptography
supabase
python-jose[cryptography]
pydantic>=2.0
python-dotenv
httpx
slowapi
